pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
urllib3 = "^1.26.12"
pandas = "^1.5.0"
bitcoin-value = "^1.5.2"
coinmarketcap-scraper = "^1.0.0"
coinmarketcap = "^5.0.3"
py-coinmarketcap-client = "^1.0.0"
requests = "^2.28.1"
py-coingecko-client = "^1.0.0"
bs4 = "^0.0.1"
matplotlib = "^3.6.1"
pandas-datareader = "^0.10.0"
mpl-finance = "^0.10.1"
seaborn = "^0.12.0"
mplfinance = "^0.12.9-beta.1"
spacy = "^3.4.2"


[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}


[tool.poetry.dependencies.en_core_web_sm]
url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1.tar.gz"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
